New 'org' django app
    move HaztrakOrg model to new 'org' django app

    move HaztrakOrg related views, services, and test to the new django 'org' app
