+++ server/apps/core/services/rcrainfo_service.py
@@ -5,7 +5,8 @@ import emanifest
 from django.db import IntegrityError
 from emanifest import RcrainfoClient, RcrainfoResponse
 
-from apps.core.models import HaztrakOrg, RcrainfoProfile
+from apps.core.models import RcrainfoProfile
+from apps.org.models import HaztrakOrg
 from apps.wasteline.models import WasteCode
 
 logger = logging.getLogger(__name__)
