+++ server/apps/core/urls.py
@@ -2,7 +2,6 @@ from dj_rest_auth.views import LoginView, LogoutView, UserDetailsView
 from django.urls import include, path
 
 from .views import (  # type: ignore
-    HaztrakOrgSitesListView,
     HaztrakProfileView,
     HaztrakSiteDetailsView,
     HaztrakSiteListView,
@@ -30,5 +29,4 @@ urlpatterns = [
     ),
     path("site", HaztrakSiteListView.as_view()),
     path("site/<str:epa_id>", HaztrakSiteDetailsView.as_view()),
-    path("org/<str:org_id>/site", HaztrakOrgSitesListView.as_view()),
 ]
