+++ server/apps/core/views/__init__.py
@@ -1,4 +1,4 @@
-from .haztrak_site import HaztrakOrgSitesListView, HaztrakSiteDetailsView, HaztrakSiteListView
+from .haztrak_site import HaztrakSiteDetailsView, HaztrakSiteListView
 from .profile_views import (
     HaztrakProfileView,
     RcrainfoProfileView,
