+++ server/apps/core/views/haztrak_site.py
@@ -2,14 +2,10 @@ import logging
 
 from django.utils.decorators import method_decorator
 from django.views.decorators.cache import cache_page
-from rest_framework import status
 from rest_framework.generics import ListAPIView, RetrieveAPIView
-from rest_framework.response import Response
-from rest_framework.views import APIView
 
 from apps.core.models import HaztrakSite
 from apps.core.serializers import HaztrakSiteSerializer
-from apps.core.services import get_org_sites
 
 logger = logging.getLogger(__name__)
 
@@ -48,13 +44,3 @@ class HaztrakSiteDetailsView(RetrieveAPIView):
             rcra_site__epa_id=epa_id, sitepermissions__profile__user=self.request.user
         )
         return queryset
-
-
-class HaztrakOrgSitesListView(APIView):
-    """Retrieve a list of sites for a given HaztrakOrg"""
-
-    @method_decorator(cache_page(60 * 15))
-    def get(self, request, *args, **kwargs):
-        haztrak_sites = get_org_sites(self.kwargs["org_id"])
-        serializer = HaztrakSiteSerializer(haztrak_sites, many=True)
-        return Response(data=serializer.data, status=status.HTTP_200_OK)
