+++ server/apps/rcrasite/admin.py
@@ -3,7 +3,7 @@ from django.urls import reverse
 from django.utils.html import format_html, urlencode
 
 from apps.core.admin import HiddenListView
-from apps.core.models import HaztrakOrg, HaztrakProfile, HaztrakSite
+from apps.core.models import HaztrakSite
 from apps.rcrasite.models import (
     Address,
     Contact,
@@ -42,27 +42,6 @@ class HaztrakSiteInline(admin.TabularInline):
 
     def has_delete_permission(self, request, obj=None):
         return False
-
-
-class HaztrakProfileInline(admin.TabularInline):
-    model = HaztrakProfile
-    extra = 0
-
-
-@admin.register(HaztrakOrg)
-class HaztrakOrgAdmin(admin.ModelAdmin):
-    list_display = ["__str__", "number_of_sites"]
-    inlines = [HaztrakSiteInline, HaztrakProfileInline]
-    readonly_fields = ["rcrainfo_integrated"]
-
-    def rcrainfo_integrated(self, obj):
-        return obj.is_rcrainfo_integrated
-
-    rcrainfo_integrated.boolean = True
-    rcrainfo_integrated.short_description = "Admin has setup RCRAInfo integration"
-
-    def number_of_sites(self, org: HaztrakOrg):
-        return HaztrakSite.objects.filter(org=org).count()
 
 
 # Register models That should only be edited within the context of another form here.
